% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_filetype_guesser.R
\name{annovar_guess_reference_table}
\alias{annovar_guess_reference_table}
\title{Guess reference table used for annovar annotation}
\usage{
annovar_guess_reference_table(path)
}
\arguments{
\item{path}{path to file}
}
\value{
one of MAF, ANNOVAR, VCF, OTHER
}
\description{
Reads the first line of a file and confirms that it is an ANNOVAR file, then returns  'ensGene' or 'refGene' 
base on whether refseq or ensemble reference tables were for gene-based annotations
}
