% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading_utils.R
\name{assert_rnaseq_df_is_formatted_correctly}
\alias{assert_rnaseq_df_is_formatted_correctly}
\title{Check RNAseq Dataframe}
\usage{
assert_rnaseq_df_is_formatted_correctly(rnaseq_df)
}
\arguments{
\item{rnaseq_df}{Dataframe containing at least three columns, named "Tumor_Sample_Barcode", "Hugo_Symbol" and "TPM". May optionally include columns named "Fold_Change" and "Transcript" (dataframe)}
}
\value{
Nothing, run for its side effects
}
\description{
Runs a bunch of assertions on the supplied rnaseq dataframe to determine whether it is valid.
Will error if it is not.
}
