% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_maf_data_pool.R
\name{assert_that_class_is_maf_data_pool}
\alias{assert_that_class_is_maf_data_pool}
\title{Assert that class is maf_data_pool}
\usage{
assert_that_class_is_maf_data_pool(object)
}
\arguments{
\item{object}{some object whose class you want to assert (anything)}
}
\value{
(invisible) TRUE if assertion succeeds, Throws an error if it doesn't
}
\description{
Assert that class is maf_data_pool object
}
\examples{
assert_that_class_is_maf_data_pool(new_maf_data_pool())
}
\seealso{
Other class_assertions: 
\code{\link{assert_that_class_is_maf_dataset_wrapper}()}
}
\concept{class_assertions}
