% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_maftools_extending_functions.R
\name{badgify}
\alias{badgify}
\title{Turn Text into A Badge}
\usage{
badgify(
  vec,
  fontweight = c("bold", "bolder", "normal", "inherit", "initial", "lighter", "unset"),
  fontsize = c("xx-small", "x-small", "small", "medium", "large", "unset", "inherit"),
  color_background = "#d2d6de",
  color_text = "#444"
)
}
\arguments{
\item{vec}{a vector of terms to make into a badge}

\item{fontweight}{fontweight of badge text}

\item{fontsize}{fontsize of badge text}

\item{color_background}{background colour. Can be a single value or a vector the same length as vec.}

\item{color_text}{text colour. Can be a single value or a vector the same length as vec.}
}
\value{
string with html encoding for the badge
}
\description{
Turn Text into A Badge
}
