% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_bool_to_colour.R
\name{bool_to_colour}
\alias{bool_to_colour}
\title{Map logical to colour}
\usage{
bool_to_colour(
  boolean,
  colour_if_true = "#69C776",
  colour_if_false = "#ED7676"
)
}
\arguments{
\item{boolean}{TRUE/FALSE (boolean)}

\item{colour_if_true}{colour to return if true (string)}

\item{colour_if_false}{colour to return if false (string)}
}
\value{
colour code (string)
}
\description{
Hex codes
}
\examples{
bool_to_colour(1==1)
}
