% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_maf_data_pool.R
\name{class_is_maf_data_pool}
\alias{class_is_maf_data_pool}
\title{Check if class is a maf_data_pool}
\usage{
class_is_maf_data_pool(object)
}
\arguments{
\item{object}{some object whose class you want to check}
}
\value{
TRUE if class of object is maf_data_pool, FALSE if not. (logical)
}
\description{
Check if class is a maf_data_pool object
}
\examples{
class_is_maf_data_pool("Hi")
class_is_maf_data_pool(new_maf_data_pool())
}
