% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_maf_dataset_wrapper.R
\name{class_is_maf_dataset_wrapper}
\alias{class_is_maf_dataset_wrapper}
\title{Check if class is a maf_dataset_wrapper}
\usage{
class_is_maf_dataset_wrapper(object)
}
\arguments{
\item{object}{some object whose class you want to check}
}
\value{
TRUE if class of object is maf_dataset_wrapper, FALSE if not. (logical)
}
\description{
Check if class is a maf_dataset_wrapper object
}
\examples{
class_is_maf_dataset_wrapper("Hi")
}
