% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{classify_url}
\alias{classify_url}
\alias{is_ad_image}
\alias{is_web_scheme}
\alias{is_likely_article}
\alias{is_likely_video}
\alias{is_likely_audio}
\alias{is_likely_binary_doc}
\alias{is_likely_archive}
\alias{is_likely_executable}
\alias{is_likely_image}
\title{Classify a URL with or without making network calls}
\usage{
classify_url(x, resolve_redirects = FALSE)

is_ad_image(x, resolve_redirects = FALSE)

is_web_scheme(x, resolve_redirects = FALSE)

is_likely_article(x, resolve_redirects = FALSE)

is_likely_video(x, resolve_redirects = FALSE)

is_likely_audio(x, resolve_redirects = FALSE)

is_likely_binary_doc(x, resolve_redirects = FALSE)

is_likely_archive(x, resolve_redirects = FALSE)

is_likely_executable(x, resolve_redirects = FALSE)

is_likely_image(x, resolve_redirects = FALSE)
}
\arguments{
\item{x}{URL to classify}

\item{resolve_redirects}{if \code{TRUE} resolve redirects such as when Facebook or
Google show an interstitial page instead of redirecting the user to
the actual URL.}
}
\description{
Classify a URL with or without making network calls
}
