% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_installation.R
\name{create_mac_binary}
\alias{create_mac_binary}
\title{#' Install dependencies
#'
#' @export
#'
install_dependencies <- function(){
  BiocManager::install("PoisonAlien/TCGAmutations")
  #BiocManager::install("maftools")
}}
\usage{
create_mac_binary()
}
\description{
#' Install dependencies
#'
#' @export
#'
install_dependencies <- function(){
  BiocManager::install("PoisonAlien/TCGAmutations")
  #BiocManager::install("maftools")
}
}
