% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRUX-package.R
\docType{package}
\name{CRUX-package}
\alias{CRUX}
\alias{CRUX-package}
\title{CRUX: Easily explore patterns of somatic variation in cancer using 'CRUX'}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Shiny app for exploring somatic variation in cancer. Powered by maftools.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/CCICB/CRUX}
  \item Report bugs at \url{https://github.com/CCICB/CRUX/issues}
}

}
\author{
\strong{Maintainer}: Sam El-Kamand \email{selkamand@ccia.org.au}

}
\keyword{internal}
