% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crux-package.R
\docType{package}
\name{crux}
\alias{crux}
\alias{crux-package}
\title{...}
\description{
\itemize{
\item URL: \url{https://gitlab.com/hrbrmstr/crux}
\item BugReports: \url{https://gitlab.com/hrbrmstr/crux/issues}
}
}
\author{
Bob Rudis (bob@rud.is)
}
