% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_download_maf.R
\name{download_maf}
\alias{download_maf}
\title{Download MAF}
\usage{
download_maf(maf, file)
}
\arguments{
\item{maf}{a loaded maf object}

\item{file}{the path to save maf to}
}
\value{
NOTHING. Run for its side effects
}
\description{
Downloads a MAF file (inc. silent mutations). Does NOT download clinical datafile.
}
\examples{
download_maf(TCGAmutations::tcga_load("GBM"), "tcga_gbm.maf")
}
