% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{external_tool_metadata}
\alias{external_tool_metadata}
\title{Prices of 50,000 round cut diamonds.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 10 rows and 13 columns.
}
\source{
external_tools_load_all_tools()
}
\usage{
external_tool_metadata
}
\description{
A dataset describing the metadata of external tools used in the cancer space.
}
\keyword{datasets}
