% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_external_tools_load_dataframe.R
\name{external_tools_add_tool_to_dataframe}
\alias{external_tools_add_tool_to_dataframe}
\title{external_tools_add_tool_to_dataframe}
\usage{
external_tools_add_tool_to_dataframe(
  external_tools_df = dplyr::tibble(),
  tool_name,
  tool_id,
  tool_group,
  tool_class,
  tool_description,
  instructions = "No instructions available yet. You're on your own buddy",
  platform = "Web App",
  website,
  doi,
  requires_maf_export = TRUE,
  requires_gene_selection = FALSE,
  maf_conversion_function = NA,
  extension = "tsv"
)
}
\arguments{
\item{external_tools_df}{the dataframe to add tool metadata to. By default, will create and return a new dataframe}

\item{tool_name}{name of tool (string)}

\item{tool_id}{id of tool (string)}

\item{tool_group}{research group that built/maintains the tool (string)}

\item{tool_class}{class of tool. Usually 'Positive Selection', 'Variant Interpretation' (string)}

\item{tool_description}{brief description of tool (string)}

\item{instructions}{brief description of how to use the tool. Any HTML tags in the string will be correctly resolved. (string)}

\item{platform}{what platform do we access the tool from. Examples include web, desktop app, cli  (string)}

\item{website}{url of tool (string)}

\item{doi}{publicatoin doi (string)}

\item{requires_maf_export}{does the tool require a maf to be exported in some other form (flag)}

\item{requires_gene_selection}{does user need to select a specific gene for export to work? (bool)}

\item{maf_conversion_function}{only relevent if requires_maf_export == true. A function that takes a maf_dataset_wrapper object (first argument), a filepath (second argument) and, if requires_gene_selection == TRUE, a gene name (third argument)  writes a file to that filepath. The idea is that said file can then be used as input to the specified tool.}

\item{extension}{what type of file is written by maf_conversion_function. default is 'tsv'. Used to appropriately name downloaded file  (string)}
}
\value{
dataframe containing external_tool_metadata
}
\description{
Appends tool metadata on the end of a dataframe
}
