% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_external_tools_load_dataframe.R
\name{external_tools_convert_maf_to_bbglab}
\alias{external_tools_convert_maf_to_bbglab}
\title{Conversion Function: MAF to BBGLab Cancer Genome Interpreter Input}
\usage{
external_tools_convert_maf_to_bbglab(maf, filepath)
}
\arguments{
\item{maf}{(MAF)}

\item{filepath}{(string)}
}
\value{

}
\description{
Convets MAF object to BBGLab Cancer Genome Interpreter Input (Genomic tabular format) then writes to a file.
Works as a maf_conversion_function. Please don't change arguments
}
