% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_external_tools_load_dataframe.R
\name{external_tools_convert_maf_to_bbglab_return_dataframe}
\alias{external_tools_convert_maf_to_bbglab_return_dataframe}
\title{Conversion Function: MAF to BBGLab Cancer Genome Interpreter Input}
\usage{
external_tools_convert_maf_to_bbglab_return_dataframe(maf)
}
\arguments{
\item{maf}{(MAF)}
}
\value{
bbglab compatible dataframe
}
\description{
Convets MAF object to BBGLab Cancer Genome Interpreter Input (Genomic tabular format) (a tsv).
}
