% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_external_tools_load_dataframe.R
\name{external_tools_convert_maf_to_cbioportal_mutation_mapper_return_dataframe}
\alias{external_tools_convert_maf_to_cbioportal_mutation_mapper_return_dataframe}
\title{MAF + Gene --> cBioPortal mutation_mapper_input}
\usage{
external_tools_convert_maf_to_cbioportal_mutation_mapper_return_dataframe(
  maf,
  gene_hugo_symbol
)
}
\arguments{
\item{maf}{a maftools maf object (maf)}

\item{gene_hugo_symbol}{hugo symbol representing the gene to export (string)}
}
\value{
dataframe in a form usable as input at https://www.cbioportal.org/mutation_mapper (data.frame)
}
\description{
Takes a MAF and converts variants hitting a particular gene to the form the cBioPortal Mutation Mapper uses as input.
}
\details{
You can find the tool at: https://www.cbioportal.org/mutation_mapper
}
