% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_external_tools_load_dataframe.R
\name{external_tools_convert_maf_to_vanilla_vcf_return_dataframe}
\alias{external_tools_convert_maf_to_vanilla_vcf_return_dataframe}
\title{Maf to VCF}
\usage{
external_tools_convert_maf_to_vanilla_vcf_return_dataframe(maf_df)
}
\arguments{
\item{maf_df}{either a maf object or a dataframe created using maftools_get_all_data}
}
\value{
data.frame with VCF columns
}
\description{
Convert a maf to a vanilla vcf (single sample)
}
