% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_external_tools_load_dataframe.R
\name{external_tools_get_property_by_tool_name}
\alias{external_tools_get_property_by_tool_name}
\title{external_tools_get_property_by_tool_name}
\usage{
external_tools_get_property_by_tool_name(
  tool_name,
  property_to_retrieve,
  external_tools_df = external_tool_metadata
)
}
\arguments{
\item{tool_name}{name of tool}

\item{property_to_retrieve}{which property to retrieve. see details for options. (string)}

\item{external_tools_df}{the dataframe to add tool metadata to. By default, will use global variable: GLOBAL_external_tools_dataframe}
}
\value{

}
\description{
external_tools_get_property_by_tool_name
}
\details{
\strong{ Options for property_to_retrieve }

\itemize{
\item tool_name
\item tool_id
\item tool_group
\item tool_class
\item tool_description
\item instructions
\item website
\item doi
\item requires_maf_export
\item requires_gene_selection
\item maf_conversion_function 
\item extension
}

See \code{?external_tools_add_tool_to_dataframe} for more info on these properties

run \code{external_tool_metadata} to see the built in tool metadata dataframe
}
\examples{
external_tools_get_property_by_tool_name(tool_name = "OncodriveFML", "website")
}
