% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_external_tools_load_dataframe.R
\name{external_tools_load_all_tools}
\alias{external_tools_load_all_tools}
\title{Load tool metadata into global variable}
\usage{
external_tools_load_all_tools()
}
\value{
external_tools_df with metadata of tool appeneded (data.frame). see \code{?external_tools_add_tool_to_dataframe} for more info on the columnss in this data.frame
}
\description{
Loads metadata for all tools, returning a dataframe.
If any of the constitutent functions are changed, run \code{external_tools_update_builtin_dataset}
}
\examples{
CRUX:::external_tools_load_all_tools()
}
