% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_external_tools_load_dataframe.R
\name{external_tools_load_bbglab_cgi}
\alias{external_tools_load_bbglab_cgi}
\title{Load tool metadata into global variable}
\usage{
external_tools_load_bbglab_cgi(external_tools_df = data.frame())
}
\value{
external_tools_df with metadata of tool appeneded (data.frame)
}
\description{
Appends  metadata for the "Cancer Genome Interpreter (cgi) tool onto external_tools_df
}
\examples{
external_tools_load_bbglab_cgi()
}
