% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_external_tools_load_dataframe.R
\name{external_tools_load_bbglab_oncodrive_fml}
\alias{external_tools_load_bbglab_oncodrive_fml}
\title{Load tool metadata into global variable}
\usage{
external_tools_load_bbglab_oncodrive_fml(external_tools_df = data.frame())
}
\arguments{
\item{external_tools_df}{the dataframe to add tool metadata to. By default, will create and return a new dataframe}
}
\value{
external_tools_df with metadata of tool appeneded (data.frame)
}
\description{
Loads metadata for the OncodriveFML tool into a global variable \strong{GLOBAL_external_tools_dataframe}
}
\examples{
CRUX:::external_tools_load_bbglab_cgi()
}
