% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_external_tools_load_dataframe.R
\name{external_tools_update_builtin_dataset}
\alias{external_tools_update_builtin_dataset}
\title{Update builtin external_tools dataset}
\usage{
external_tools_update_builtin_dataset()
}
\description{
Takes the output of external_tools_load_all_tools and saves it to CRUX/data as the dataset: \strong{CRUX::external_tool_metadata}.
This saved dataset is what is used by the app. If you make change to any external_tools_load_... function, you must rerun this function.
}
