% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_rendered_plot_height}
\alias{get_rendered_plot_height}
\title{Get Width/Height of a rendered element}
\usage{
get_rendered_plot_height(output_id, session, return_inches = F, dpi = 70)
}
\arguments{
\item{output_id}{"ID of the plot/element who's width/height you want to know"}

\item{session}{"Current Session. Used to get namespacing and clientData information"}

\item{return_inches}{"Should we return width/height in pixels or inches"}

\item{dpi}{"dpi used to convert pixels to inches"}
}
\value{
"Width/Height in pixels/inches"
}
\description{
Get Width/Height of a rendered element
}
