% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_load_tcga.R
\name{get_tcga_mafs}
\alias{get_tcga_mafs}
\title{Get TCGA MAFs}
\usage{
get_tcga_mafs(study_name)
}
\arguments{
\item{study_name}{Abbreviation of TCGA dataset to install. To see options, run maftools::tcgaAvailable()}
}
\value{
A single MAF object.
}
\description{
Get TCGA MAFs
}
