% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_gistic.R
\name{gistic_tabulate_amp_or_del_genes_file}
\alias{gistic_tabulate_amp_or_del_genes_file}
\title{Tabulate Gistic Gene Files}
\usage{
gistic_tabulate_amp_or_del_genes_file(path_to_amp_or_del_genes_file)
}
\arguments{
\item{path_to_amp_or_del_genes_file}{path to either del_genes.conf_XX.txt OR amp_genes.conf_90.txt. These files can be produced using gistic (string)}
}
\value{
a dataframe containing cytobands, q values, peak locations and which genes are in each peak
}
\description{
Parses a gistic path_to_amp_or_del_genes_file
}
