% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_filetype_guesser.R
\name{guess_genomic_mutation_filetype}
\alias{guess_genomic_mutation_filetype}
\title{Guesses mutation filetype}
\usage{
guess_genomic_mutation_filetype(path)
}
\arguments{
\item{path}{path to file}
}
\value{
one of MAF, ANNOVAR, VCF, OTHER
}
\description{
Reads the first line of a file and attempt to determine the filetype based on the header
}
