% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_centered_down_arrow.R
\name{html_alert}
\alias{html_alert}
\title{HTML_alert}
\usage{
html_alert(text, status = "warning")
}
\arguments{
\item{text}{Alert text (string)}

\item{status}{Boostrap status. One of primary,secondary,success,danger,warning,info,light,dark (string)}
}
\value{
html-flagged text (string")
}
\description{
HTML_alert
}
