% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_centered_down_arrow.R
\name{icon_down_arrow}
\alias{icon_down_arrow}
\title{icon_down_arrow}
\usage{
icon_down_arrow(fontsize = "40px", alignment = "center", break_after = FALSE)
}
\arguments{
\item{fontsize}{a valid css fontsize, e.g. 60px (string)}

\item{alignment}{One of 'left', 'right', 'center' or 'inherit'(string)}

\item{break_after}{insert a linebreak after icon (flag)}
}
\value{
centered down arrow fluidrow()
}
\description{
icon_down_arrow
}
\concept{arrows}
