% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_validating_clinical_feature_files.R
\name{is_valid_clinicalfeaturefile}
\alias{is_valid_clinicalfeaturefile}
\title{Is a clinical_feature_file valid for a given maf}
\usage{
is_valid_clinicalfeaturefile(clinicalData, maf)
}
\arguments{
\item{clinicalData}{Clinical data associated with each sample/Tumor_Sample_Barcode in MAF. Could be a text file or a data.frame. Default NULL.}

\item{maf}{tab delimited MAF file. File can also be gz compressed. Required. Alternatively, you can also provide already read MAF file as a dataframe.}
}
\value{
True / False (boolean)
}
\description{
Runs read.maf with a given clinicalData and maf object pair and returns TRUE ONLY if the clinicalData file is appropriate for the given MAF.
If maf object is not valid, this function will also return FALSE.
By using clinicalData=NULL the function allows testing validity of a MAF file in Isolation.
}
\details{
If clinical feature file has duplicated Tumor_Sample_Barcodes, only the first entry is considered and the file is still 'valid'.
}
