% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_links.R
\name{link}
\alias{link}
\title{Link}
\usage{
link(url, newtab = TRUE, text)
}
\arguments{
\item{url}{URL}

\item{newtab}{shoule the link be opened in a new tab? (flag)}

\item{text}{display text for link}
}
\value{
list() with a shiny.tag class. Can convert to HTML string via as.character()
}
\description{
Link
}
