% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_maf_data_pool.R
\name{maf_data_pool_add_dataset}
\alias{maf_data_pool_add_dataset}
\title{Add to MAF Data Pool}
\usage{
maf_data_pool_add_dataset(maf_dataset_wrapper, maf_data_pool)
}
\arguments{
\item{maf_dataset_wrapper}{object to add to the data pool (maf_dataset_wrapper)}

\item{maf_data_pool}{the data pool to add the dataset wrapper to (maf_data_pool)}
}
\value{
a new maf_data_pool object with the additional objects appended (maf_data_pool)
}
\description{
Add a maf_dataset_wrapper object to the maf_data_pool
}
