% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_maf_data_pool.R
\name{maf_data_pool_get_all_nonfunction_property_names}
\alias{maf_data_pool_get_all_nonfunction_property_names}
\title{Get Colnames for maf_data_pool to data.frame conversion}
\usage{
maf_data_pool_get_all_nonfunction_property_names(maf_data_pool)
}
\arguments{
\item{maf_data_pool}{data pool of interest (maf_data_pool)}
}
\value{
a vector of list_within_list 'properties' that are not functions (character)
}
\description{
Gets the names of all maf_dataframe_wrapper properties that don't hold functions. 
This information is useful when converting to a data.frame which can't hold functions
}
\details{
This function works by only looking at the first element in the list, which means if one day we change the constructor such that not all wrappers have the same set of properties, we may see unexpected behaviour.
Since it is extremely unlikely this will change, I won't make this more robust for now. If I am wrong ... sorry future me :(
}
