% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_maf_data_pool.R
\name{maf_data_pool_get_index_from_unique_name}
\alias{maf_data_pool_get_index_from_unique_name}
\title{Retrieve dataset wrapper from data pool using unique_name}
\usage{
maf_data_pool_get_index_from_unique_name(maf_data_pool, unique_name)
}
\arguments{
\item{maf_data_pool}{The data pool of interest  (maf_data_pool)}

\item{unique_name}{unique_name of the maf_dataset_wrapper you're interested in(string).}
}
\value{
index of the specified maf_dataset_wrapper within the maf_data_pool
}
\description{
Will throw error if unique_name is not found
}
