% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading_utils.R
\name{maf_data_pool_load_data}
\alias{maf_data_pool_load_data}
\title{Load a dataset wrapper from data pool using unique_name}
\usage{
maf_data_pool_load_data(maf_data_pool, unique_name)
}
\arguments{
\item{maf_data_pool}{The data pool of interest  (maf_data_pool)}

\item{unique_name}{unique_name of the maf_dataset_wrapper you want to load (string).}
}
\value{
maf_data_pool object with the wrapped dataset now loaded (maf_data_pool)
}
\description{
use maf_data_pool_robust_load instead. It wraps this function
Will throw error if unique_name is not found or status is anything other than 'not_loaded'.
}
