% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_maf_data_pool.R
\name{maf_data_pool_make_name_unique}
\alias{maf_data_pool_make_name_unique}
\title{Make a name unique (within datapool)}
\usage{
maf_data_pool_make_name_unique(
  maf_data_pool,
  name,
  max_number_of_attempts = 50
)
}
\arguments{
\item{maf_data_pool}{The data pool of interest  (maf_data_pool)}

\item{name}{name to make unique (string)}

\item{max_number_of_attempts}{max number of attempts (whole number)}
}
\value{
a unique name (string)
}
\description{
Checks if a name is unique in a given data pool. 
If not, the function will append 15 letters randomly to the name until it finds a unique name, or surpasses \code{max_number_of_attempts}. 
Will throw an error if it can't find a unique name
}
