% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_maf_dataset_wrapper.R
\name{maf_data_pool_robust_load}
\alias{maf_data_pool_robust_load}
\title{Load MAF objects}
\usage{
maf_data_pool_robust_load(maf_data_pool, unique_name)
}
\arguments{
\item{maf_data_pool}{The data pool of interest  (maf_data_pool)}

\item{unique_name}{unique_name of the maf_dataset_wrapper you want to load (string).}
}
\value{
maf_data_pool with the specified datawrapper loaded (maf_data_pool)
}
\description{
This function allows user to pass a maf data pool and a unique name of the dataset of interest.
It differs from maf_data_pool_load_data in that maf_data_pool_load_data will throw an error if the status of the loaded object is anything other than "not_loaded".
This function will take any valid status and try to get the data loaded.
It decides if the data actually needs loading, and if so, loads it.

It is designed to run before you run maf_data_pool_unique_name_to_maf / maf_data_pool_get_data_wrapper_from_unique_name.
}
