% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_maf_data_pool.R
\name{maf_data_pool_to_dataframe}
\alias{maf_data_pool_to_dataframe}
\title{MAF Data Pool to Dataframe}
\usage{
maf_data_pool_to_dataframe(maf_data_pool)
}
\arguments{
\item{maf_data_pool}{the data pool to convert to a dataframe (maf_data_pool)}
}
\value{
a data.frame containing all properties of maf_data_pool except for those containing functions (data.frame)
}
\description{
Converts a data pool to a data.frame form.
}
