% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise_data_pool.R
\name{maf_data_pool_too_network_map}
\alias{maf_data_pool_too_network_map}
\title{Network Map View of Data Pool}
\usage{
maf_data_pool_too_network_map(
  maf_data_pool,
  size_of_sources = 10,
  size_of_datasets = 7
)
}
\arguments{
\item{maf_data_pool}{data pool to visualise (maf_data_pool)}
}
\value{
sigmaJS htmlwidget
}
\description{
Takes a maf_data_pool and visualizes it as a of a sigmaJS network
}
\examples{
maf_data_pool_too_network_map(pcawg_datasets_to_data_pool(maf_data_pool = new_maf_data_pool()))
}
