% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_maf_dataset_wrapper.R
\name{maf_data_pool_unique_name_to_maf_nonreactive}
\alias{maf_data_pool_unique_name_to_maf_nonreactive}
\title{maf_data_pool_unique_name_to_maf_nonreactive}
\usage{
maf_data_pool_unique_name_to_maf_nonreactive(maf_data_pool, unique_name)
}
\arguments{
\item{maf_data_pool}{the data pool to add the dataset wrapper to (maf_data_pool)}

\item{unique_name}{unique name (string)}
}
\value{
return loaded data (maf object, nonreactive)
}
\description{
Takes a unique name and returns the maf object associated with said entry in the maf_data_pool. 
Involves loading the dataset but does NOT actually update the maf_data_pool.
}
