% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_maf_dataset_wrapper.R
\name{maf_data_pool_unique_name_to_maf_reactive}
\alias{maf_data_pool_unique_name_to_maf_reactive}
\title{maf_data_pool_unique_name_to_maf_reactive}
\usage{
maf_data_pool_unique_name_to_maf_reactive(maf_data_pool, unique_name)
}
\arguments{
\item{maf_data_pool}{the reactiveVal we want will search for our dataset of interest, and is also the object we apply any changes to. (reactiveVal)}

\item{unique_name}{unique_name of the maf_dataset_wrapper fetch (string).}
}
\value{
maf object (maf)
}
\description{
Loads the relevant MAF if possible, appplies changes to maf_data_pool reactive then returns the loaded maf
}
