% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading_utils.R
\name{maf_data_pool_unload_data}
\alias{maf_data_pool_unload_data}
\title{Unload dataset wrapper from data pool using unique_name}
\usage{
maf_data_pool_unload_data(maf_data_pool, unique_name)
}
\arguments{
\item{maf_data_pool}{The data pool of interest  (maf_data_pool)}

\item{unique_name}{unique_name of the maf_dataset_wrapper you want to unload (string).}
}
\value{
maf_data_pool object with the wrapped dataset now unloaded (maf_data_pool)
}
\description{
Will throw error if unique_name is not found
}
