% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading_utils.R
\name{maf_data_set_wrapper_load_data}
\alias{maf_data_set_wrapper_load_data}
\title{Load data into memory}
\usage{
maf_data_set_wrapper_load_data(maf_dataset_wrapper)
}
\arguments{
\item{maf_dataset_wrapper}{the wrapper of the dataset you want to load into memory (maf_dataset_wrapper)}
}
\value{
a copy of the original wrapper with status and loaded data updated. (maf_dataset_wrapper)
}
\description{
Uses the functions in maf_dataset_wrapper to load data into the loaded_data element and update status.
}
\details{
It is only the maf object that is loaded. RNA data stays sitting in a file until it is retrieved via maf_dataset_wrapper_get_rnaseq
}
\examples{
#Generate wrapper
my_data <- CRUX:::tcga_dataset_to_maf_dataset_wrapper(maf_data_pool = CRUX:::new_maf_data_pool(), tcga_study_abbreviation = "ACC")

#Load data
my_data <- CRUX:::maf_data_set_wrapper_load_data(my_data)

#Access loaded data
print(my_data$loaded_data)

#Unload when finished
my_data <- CRUX:::maf_data_set_wrapper_unload_data(my_data)

}
\seealso{
Other data_set_wrapper_loading: 
\code{\link{maf_data_set_wrapper_unload_data}()}
}
\concept{data_set_wrapper_loading}
