% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading_utils.R
\name{maf_data_set_wrapper_unload_data}
\alias{maf_data_set_wrapper_unload_data}
\title{Unload data from memory}
\usage{
maf_data_set_wrapper_unload_data(maf_dataset_wrapper)
}
\arguments{
\item{maf_dataset_wrapper}{the wrapper of the dataset you want to unload into memory (maf_dataset_wrapper)}
}
\value{
a copy of the original wrapper with status and loaded data elements updated. Status is changed from "ready" => "not_loaded". loaded_data is changed from maf_object to NA. (maf_dataset_wrapper)
}
\description{
Uses the functions in maf_dataset_wrapper to load data into the loaded_data element and update status.
}
\examples{
#Generate wrapper
my_data <- tcga_dataset_to_maf_dataset_wrapper(tcga_study_abbreviation = "ACC")

#Load data
my_data <- maf_data_set_wrapper_load_data(my_data)

#Access loaded data
print(my_data$loaded_data)

#Unload when finished
my_data <- maf_data_set_wrapper_unload_data(my_data)
}
\seealso{
Other data_set_wrapper_loading: 
\code{\link{maf_data_set_wrapper_load_data}()}
}
\concept{data_set_wrapper_loading}
