% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading_utils.R
\name{maf_data_wrapper_get_rnaseq_data_for_samples_with_mutation_data}
\alias{maf_data_wrapper_get_rnaseq_data_for_samples_with_mutation_data}
\title{Get Expression Data from maf}
\usage{
maf_data_wrapper_get_rnaseq_data_for_samples_with_mutation_data(
  maf_dataset_wrapper
)
}
\arguments{
\item{maf_dataset_wrapper}{a maf_dataset wrapper. See ?new_maf_dataset_wrapper for details.}
}
\value{
RNAseq data if present or NULL if no RNA data has been associated with it yet. Will only return expression data for samples that have corresponding mutation data (dataframe)
}
\description{
Similar tomaf_data_wrapper_get_rnaseq_df but only returns expression data for samples that have corresponding mutation data
}
\examples{
# Prepare Data
rna_path = system.file("example_data/blca_rnaseq.tsv", package = "CRUX")
maf_data_wrapper = tcga_dataset_to_maf_dataset_wrapper(new_maf_data_pool(), "BLCA")

# Add RNA data
maf_data_wrapper_with_RNA = maf_data_wrapper_add_rnaseq(maf_data_wrapper, rnaseq_path = rna_path) 

# Retrieve RNA data for samples with mutation data
maf_data_wrapper_get_rnaseq_data_for_samples_with_mutation_data(maf_data_wrapper_with_RNA)

}
