% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading_utils.R
\name{maf_data_wrapper_get_rnaseq_df}
\alias{maf_data_wrapper_get_rnaseq_df}
\title{Get rnaseq data}
\usage{
maf_data_wrapper_get_rnaseq_df(maf_dataset_wrapper)
}
\arguments{
\item{maf_dataset_wrapper}{a maf_dataset wrapper. See ?new_maf_dataset_wrapper for details.}
}
\value{
RNAseq data if present or NULL if no RNA data has been associated with it yet (dataframe)
}
\description{
You probaly want maf_data_wrapper_get_rnaseq_data_for_samples_with_mutation_data instead.
}
\details{
This function takes a maf_dataset_wrapper and retrieves ALL rnaseq data present (inc. for samples that would have been filtered out by subsetting operations).
}
