% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading_utils.R
\name{maf_data_wrapper_has_rnaseq_data}
\alias{maf_data_wrapper_has_rnaseq_data}
\title{Maf Dataset has RNAseq data}
\usage{
maf_data_wrapper_has_rnaseq_data(maf_dataset_wrapper)
}
\arguments{
\item{maf_dataset_wrapper}{a maf dataset wrapper. See ?new_maf_dataset_wrapper}
}
\value{
TRUE if RNAseq data is present, False if not (Boolean)
}
\description{
Maf Dataset has RNAseq data
}
