% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_maftools_fix_clinical_data_types.R
\name{maftools_add_brca_subtype}
\alias{maftools_add_brca_subtype}
\title{add BRCA subtype clinical feature}
\usage{
maftools_add_brca_subtype(maf)
}
\arguments{
\item{maf}{maftools maf object}
}
\value{
maf object
}
\description{
Checks if maf is BRCA and contains cols required to classify as Triple Negative / Not Triple Negative.
If so, adds a 'subtype' column describing whether its triple negative or not triple negative
}
