% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_maftools_extending_functions.R
\name{maftools_add_clinical_data}
\alias{maftools_add_clinical_data}
\title{Add clindata to MAF}
\usage{
maftools_add_clinical_data(maf, clindata_path)
}
\arguments{
\item{maf}{an existing MAF object as produced by maftools::read.maf (MAF)}

\item{clindata_path}{Eithera path to a csv/tsv that contains sample level metadata. Must include a}
}
\value{
maf object with
}
\description{
Adds 'extra' clinical metadata to a maf object with existing metadata
}
