% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_maftools_extending_functions.R
\name{maftools_check_clinical_feature_oncoplot_compatibility}
\alias{maftools_check_clinical_feature_oncoplot_compatibility}
\title{Check if clinical features are plottable on an Oncoplot}
\usage{
maftools_check_clinical_feature_oncoplot_compatibility(maf, clinFeature, ...)
}
\arguments{
\item{clinical_annotation}{a vector that represents a value of clinical annotations}
}
\value{
String describing reason column can't be plotted

String describing reason column can't be plotted
}
\description{
Check if clinical features are plottable on an Oncoplot. See details for the rules used
}
\details{
Rules:

1. If column is non-nummeric and contains >100 unique elements maftools 'get_anno_cols' throws the error 'Error in sample.int(length(x), size, replace, prob) : 
cannot take a sample larger than the population when "replace = FALSE"'


Check if clinical features are plottable on an Oncoplot

Check if clinical features are plottable on an Oncoplot. See details for the rules used

Rules:

1. If column is non-nummeric and contains >100 unique elements maftools 'get_anno_cols' throws the error 'Error in sample.int(length(x), size, replace, prob) : 
cannot take a sample larger than the population when "replace = FALSE"'
}
