% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_maftools_extending_functions.R
\name{maftools_clinical_data_get_levels}
\alias{maftools_clinical_data_get_levels}
\title{maftools_clinical_data_get_levels}
\usage{
maftools_clinical_data_get_levels(maf, clinical_feature)
}
\arguments{
\item{maf}{}

\item{clinical_feature}{}
}
\value{
Number of distinct levels of a clinical feature
}
\description{
maftools_clinical_data_get_levels
}
