% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_maftools_extending_functions.R
\name{maftools_clinical_data_visually_summarise}
\alias{maftools_clinical_data_visually_summarise}
\title{Plot Clinical Data
Plot a particular column of clinical DATA.}
\usage{
maftools_clinical_data_visually_summarise(
  maf,
  clinical_feature = "Tumor_Sample_Barcode",
  threshold = NULL,
  selected_items = NULL,
  distance_from_bar = 2
)
}
\arguments{
\item{maf}{MAF object (MAF)}

\item{clinical_feature}{Name of a clinical feature (string)}
}
\value{
ggplot / grob
}
\description{
Plot Clinical Data
Plot a particular column of clinical DATA.
}
