% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_maftools_extending_functions.R
\name{maftools_clinical_feature_description}
\alias{maftools_clinical_feature_description}
\title{Get a data.frame describing each clinical feature of a MAF}
\usage{
maftools_clinical_feature_description(
  maf,
  checkmark = c("none", "oncoplottable")
)
}
\arguments{
\item{maf}{a maftools maf object}

\item{checkmark}{should a pass/fail checkmark icon be added before the annotations name. (none = no pass/fail checkmark; oncoplottable = whether feature can be added to an oncoplot)}
}
\value{
a data.frame with 3 columns. annotation (feature name); type (whether feature is numeric/categorical); levels (number of distinct, non-na values); content (html string with names + badges with additional information)
}
\description{
Get a data.frame describing each clinical feature of a MAF
}
\details{
Rule for pass/fail checks: 
To be 'oncoplottable' there must be <= 100 non-missing levels OR the feature must be numeric
}
