% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_maftools_extending_functions.R
\name{maftools_cluster_samples}
\alias{maftools_cluster_samples}
\title{maftools_cluster_samples}
\usage{
maftools_cluster_samples(
  maf,
  custom_genelist_to_cluster_by = NULL,
  number_of_genes = 50,
  genes_to_annotate = NULL,
  annotate_most_altered_genes = TRUE,
  topn_genes = 5,
  metadata_columns = NULL,
  include_silent_mutations = FALSE,
  show_rownames = FALSE,
  show_colnames = FALSE,
  annotation_legend = TRUE,
  fontsize = 10
)
}
\arguments{
\item{maf}{maf object from maftools package (maf)}

\item{custom_genelist_to_cluster_by}{names of genes to base clustering on. If unsure what genes to use. By default, the top 50 genes ranked by how manys samples they are mutated in will be used (character vector)}

\item{number_of_genes}{number of genes to cluster based on. Chooses genes which are mutated in the most samples  (only matters if not supplying \strong{custom_genelist_to_cluster_by}) (integer)}

\item{genes_to_annotate}{a custom list of genes to plot as a pseudo oncoplot (string)}

\item{annotate_most_altered_genes}{automatically annotate with mutational status of genes mutated in the most samples (this is always based on coding/splice site mutation) (bool)}

\item{topn_genes}{if annotate_most_altered_genes is true, how many genes to automatically visualise (integer)}

\item{metadata_columns}{name of metadata columns to annotate heatmap based on (character)}

\item{include_silent_mutations}{consider a gene mutated even if the only mutations present are silent (bool)}

\item{show_rownames}{show sample names on rows (bool)}

\item{show_colnames}{show sample names on columns  (bool)}
}
\value{
pheatmap object
}
\description{
Identify and visualise clusters of cancer samples based on somatic mutation data (gene-level differences visualised, not variant-level)
 
 Approach involves:
 \enumerate{
 \item Selecting a geneset of interest (by default uses genes mutated in the most samples).
 \item Calculating \strong{1-jaccard} distance between samples based on which genes of the genesets are mutated.
 \item Running heirarchical clustering algorithm using \strong{pheatmap}.
 \item Visualise resulting heatmap with user-selected annotations.
 }
}
\examples{
maf <- TCGAmutations::tcga_load("GBM", source = "Firehose") 
maftools_cluster_samples(maf)
}
