% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_maftools_fix_clinical_data_types.R
\name{maftools_fix_clinical_data_types}
\alias{maftools_fix_clinical_data_types}
\title{Type MAF clinical data}
\usage{
maftools_fix_clinical_data_types(maf)
}
\arguments{
\item{maf}{a maf object (MAF)}
}
\value{
MAF object with correctly typed clinical.data dataframe (MAF)
}
\description{
Returns a MAF with correctly typed clinical data. 
By default, many maftools operations (e.g. subsetting) results in a loss of clinical data type (everything gets cast as characters).
This function can be run on any MAF to resolve this issue.
}
\examples{
maf = TCGAmutations::tcga_load(study = "GBM", source = "Firehose")
maftools::getClinicalData(maf) \%>\% str()
new_maf = maftools_fix_clinical_data_types(maf)
maftools::getClinicalData(new_maf) \%>\% str()
}
