% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_maftools_extending_functions.R
\name{maftools_get_all_data}
\alias{maftools_get_all_data}
\title{Get MAF data}
\usage{
maftools_get_all_data(maf, include_silent_mutations = T)
}
\arguments{
\item{maf}{}

\item{include_silent_mutations}{}
}
\value{
a data.frame in MAF form where each variant has a separate row (data.frame)
}
\description{
Input a maf object. Returns table containing ALL MAF data. Nonsynonymous AND synonymous
}
\examples{
maftools_get_all_data(TCGAmutations::tcga_load("GBM"))
}
