% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_external_tools_load_dataframe.R
\name{maftools_get_transcript_refseq}
\alias{maftools_get_transcript_refseq}
\title{Get Longest Transcript Refseq ID}
\usage{
maftools_get_transcript_refseq(maf_df)
}
\arguments{
\item{maf_df}{(HGNC ID of gene)}
}
\value{
vector of the same length as maf_df containing transcript refseq ids (character)
}
\description{
Returns the transcript id from the 'Transcript_ID' column of a maf_df.
If no Transcript_ID column is found, will return the longest refseq transcript for each gene name.
}
