% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_maftools_extending_functions.R
\name{maftools_number_of_samples}
\alias{maftools_number_of_samples}
\title{Number of samples in a MAF}
\usage{
maftools_number_of_samples(maf)
}
\arguments{
\item{maf}{@inherit maftools::read.maf}
}
\value{
int
}
\description{
Number of samples in a MAF
}
\examples{
CRUX:::maftools_number_of_samples(maftools::tcgaLoad("GBM"))
}
