% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_maftools_get_all_data.R
\name{maftools_plot_rainfall}
\alias{maftools_plot_rainfall}
\title{Maftools rainfallPlot Wrapper}
\usage{
maftools_plot_rainfall(
  maf,
  tsb,
  detectChangePoints = TRUE,
  ref.build = "hg19",
  pointSize = 0.4,
  fontSize = 1.2
)
}
\arguments{
\item{maf}{an \code{\link[maftools]{MAF}} object generated by \code{\link[maftools]{read.maf}}. Required.}

\item{tsb}{specify sample names (Tumor_Sample_Barcodes) (string)}

\item{detectChangePoints}{If TRUE, detectes genomic change points where potential kataegis are formed. Results are written to an output tab delimted file.}

\item{ref.build}{Reference build for chromosome sizes. Can be hg18, hg19 or hg38. Default hg19.}

\item{pointSize}{Default 0.8.}

\item{fontSize}{Default 12.}
}
\value{
predicted kataegis sites (dataframe)
}
\description{
Wraps maftools::rainfallPlot. The original function plots a rainfall graph, prints the predicted kataegis coords as a df, and writes it to a file.
This function does all of this EXCEPT it hides the printed df, reads the df from the file and returns it. It also deletes the created file.
}
