% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_maftools_extending_functions.R
\name{maftools_remove_dubious_genes}
\alias{maftools_remove_dubious_genes}
\title{Filter dubious genes from MAF}
\usage{
maftools_remove_dubious_genes(maf, genelist = somaticflags::somaticflags)
}
\arguments{
\item{maf}{a MAF object}

\item{genelist}{a character vector containing HUGO Symbols to remove from MAF}
}
\value{
MAF object
}
\description{
Filter out genes likely to appear in many analyses from your MAF (e.g. TTN & Olfactory receptors )
}
\examples{
maftools_filter_out_dubious_genes(TCGAmutations::tcga_load("gbm"), "TTN")
}
