% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_data_import_maf_path_to_maf.R
\name{mod_data_import_maf_path_to_maf_server}
\alias{mod_data_import_maf_path_to_maf_server}
\title{mod_data_import_step1 Server Functions}
\usage{
mod_data_import_maf_path_to_maf_server(id, maf_path, clinicalData)
}
\arguments{
\item{maf_path}{Path to maf file  (string) (reactive)}

\item{clinicalData}{Clinical data associated with each sample/Tumor_Sample_Barcode in MAF. Could be a text file or a data.frame. Default NULL.}
}
\value{
maf object if read was successful. NULL if it was not (MAF) (reactive)
}
\description{
mod_data_import_step1 Server Functions
}
