% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_data_import_maf_path_to_maf.R
\name{mod_data_import_maf_path_to_maf_ui}
\alias{mod_data_import_maf_path_to_maf_ui}
\title{mod_data_import_step1 UI Function}
\usage{
mod_data_import_maf_path_to_maf_ui(id)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}
}
\description{
GUI for taking a filepath returning a maf object.
Includes a panel containing the MAF file summary or an error message if the supplied path does not point to a valid maf.
}
