% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_data_import_step2.R
\name{mod_data_import_step2_server}
\alias{mod_data_import_step2_server}
\title{data_import_step_2 Server Functions}
\usage{
mod_data_import_step2_server(
  id,
  default_data_source = reactive(NULL),
  default_display_name = reactive(NULL),
  default_short_name = reactive(NULL),
  default_description = reactive(NULL)
)
}
\value{
named list with elements: 
\enumerate{
\item all_valid (bool)
\item display_name (string)
\item short_name (string)
\item data_source (string)
\item description (string)
}
}
\description{
Returns a list of user-filled metadata. Do \strong{NOT} evaluate any of the properties before checking all_valid == TRUE (list) (reactive).
}
\details{
The list itself is reactive, so to access elements do:

\preformatted{
metadata <- mod_data_import_step2_server(id="my_id")

# In a reactive context:
observe({
    metadata()$all_valid
    metadata()$display_name
    metadata()$short_name
    metadata()$data_source
    metadata()$description
})
}
}
