% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_external_tools.R
\name{mod_external_tools_server}
\alias{mod_external_tools_server}
\title{external_tools Server Functions}
\usage{
mod_external_tools_server(id, maf_data_pool)
}
\arguments{
\item{id}{Internal parameters for {shiny}.}

\item{maf_data_pool}{the data pool to add the dataset wrapper to (maf_data_pool)}
}
\description{
external_tools Server Functions
}
