% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_import_clinical_featurefile.R
\name{mod_import_clinical_featurefile_ui}
\alias{mod_import_clinical_featurefile_ui}
\title{import_clinical_featurefile UI Function}
\usage{
mod_import_clinical_featurefile_ui(id)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}
}
\description{
A shiny Module.
}
