% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_plot_apobec_diff.R
\name{mod_plot_apobec_diff_server}
\alias{mod_plot_apobec_diff_server}
\title{plot_apobec_diff Server Functions}
\usage{
mod_plot_apobec_diff_server(id, maf, tnm)
}
\arguments{
\item{maf}{an \code{\link[maftools]{MAF}} object used to generate the matrix}

\item{tnm}{output generated by \code{\link[maftools]{trinucleotideMatrix}}}
}
\description{
plot_apobec_diff Server Functions
}
