% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_plot_somatic_interactions.R
\name{mod_plot_somatic_interactions_server}
\alias{mod_plot_somatic_interactions_server}
\title{somatic_interactions Server Functions}
\usage{
mod_plot_somatic_interactions_server(id, maf)
}
\arguments{
\item{maf}{a maf object (maf)}
}
\description{
somatic_interactions Server Functions
}
