% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_render_clinical_data_table.R
\name{mod_render_clinical_data_table_server}
\alias{mod_render_clinical_data_table_server}
\title{Title}
\usage{
mod_render_clinical_data_table_server(id, maf)
}
\arguments{
\item{maf}{maf object (reactive)}
}
\value{
Nothing. Run for its side effects
}
\description{
Title
}
