% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_render_clinical_data_table.R
\name{mod_render_clinical_data_table_ui}
\alias{mod_render_clinical_data_table_ui}
\title{render_clinical_data_table UI Function}
\usage{
mod_render_clinical_data_table_ui(id)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}
}
\description{
A shiny Module. Renders clinica data attached to a maf object in the form of a datatable
}
