% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_render_downloadable_dataframe.R
\name{mod_render_downloadabledataframe_server}
\alias{mod_render_downloadabledataframe_server}
\title{Title}
\usage{
mod_render_downloadabledataframe_server(
  id,
  tabular_data_object,
  basename,
  rownames = FALSE,
  colnames = TRUE,
  filter = "top",
  message_if_tabular_data_is_null = "Please select valid mutalisk files"
)
}
\arguments{
\item{id}{shiny paramater}

\item{tabular_data_object}{tabular data object (usually data.frame or data.table)}

\item{basename}{name of downloaded file (flag)}

\item{rownames}{download with rownames (flag)}

\item{colnames}{download with colnames (flag)}

\item{filter}{Position of filter search box: one of 'top', 'bottom' or 'none'  (string)}
}
\description{
Title
}
