% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/mod_select_dataset_from_maf_data_pool_pickerinput_and_return_maf_dataset_wrapper.R
\name{mod_select_dataset_from_maf_data_pool_pickerinput_and_return_maf_dataset_wrapper_ui}
\alias{mod_select_dataset_from_maf_data_pool_pickerinput_and_return_maf_dataset_wrapper_ui}
\title{select_dataset_from_maf_data_pool_pickerinput_and_return_maf_and_metadata UI Function}
\usage{
mod_select_dataset_from_maf_data_pool_pickerinput_and_return_maf_dataset_wrapper_ui(
  id,
  panel = TRUE
)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}

\item{panel}{should pickerinput be wrapped in a shinyWidgets::panel (flag)}
}
\description{
A shiny Module.
}
