% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/mod_select_dataset_from_maf_data_pool_pickerinput_and_return_maf.R
\name{mod_select_dataset_from_maf_data_pool_pickerinput_and_return_maf_ui}
\alias{mod_select_dataset_from_maf_data_pool_pickerinput_and_return_maf_ui}
\title{mod_select_dataset_from_data_pool_dropdown_and_return_maf_}
\usage{
mod_select_dataset_from_maf_data_pool_pickerinput_and_return_maf_ui(
  id,
  panel = TRUE
)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}

\item{panel}{should pickerinput be wrapped in a shinyWidgets::panel (flag)}
}
\description{
UI for selecting a dataset from the maf_data_pool.
}
