% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/mod_select_dataset_from_maf_data_pool_pickerinput.R
\name{mod_select_dataset_from_maf_data_pool_pickerinput_server}
\alias{mod_select_dataset_from_maf_data_pool_pickerinput_server}
\title{select_dataset_from_maf_data_pool_pickerinput Server Functions}
\usage{
mod_select_dataset_from_maf_data_pool_pickerinput_server(
  id,
  maf_data_pool,
  label = "Dataset",
  style = "btn-outline-primary",
  multiple = FALSE,
  max_selected_datasets = 40
)
}
\arguments{
\item{style}{When set to a string, add the value to the button's style. Type: string | null; Default: null.}
}
\value{
a character vector of unique_names or NULL if none are selected
}
\description{
select_dataset_from_maf_data_pool_pickerinput Server Functions
}
