% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_select_dataset_from_maf_data_pool.R
\name{mod_select_datasets_from_maf_data_pool_server}
\alias{mod_select_datasets_from_maf_data_pool_server}
\title{select_dataset_from_maf_data_pool Server Functions}
\usage{
mod_select_datasets_from_maf_data_pool_server(
  id,
  maf_data_pool,
  max_selected_datasets = -1
)
}
\arguments{
\item{maf_data_pool}{the maf_data_pool (maf_data_pool; reactiveVal)}

\item{max_selected_datasets}{max number of items that can be selected. -1 means unlimited (int)}
}
\value{
a character vector containing the unique_names of all datasets selected or NULL (NULL / character)
}
\description{
select_dataset_from_maf_data_pool Server Functions
}
