% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_select_dataset_from_maf_data_pool.R
\name{mod_select_datasets_from_maf_data_pool_ui}
\alias{mod_select_datasets_from_maf_data_pool_ui}
\title{select_dataset_from_maf_data_pool UI Function}
\usage{
mod_select_datasets_from_maf_data_pool_ui(id, label = "select")
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}

\item{label}{"a"}
}
\description{
A shiny Module.
}
