% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/mod_select_dataset_from_maf_data_pool_pickerinput_and_return_maf_dataset_wrapper.R
\name{mod_select_maf_dataset_wrapper_server}
\alias{mod_select_maf_dataset_wrapper_server}
\title{Select Dataset, Return maf_dataset_wrapper}
\usage{
mod_select_maf_dataset_wrapper_server(id, maf_data_pool, label = "Dataset")
}
\arguments{
\item{id}{Internal parameters for {shiny}.}

\item{maf_data_pool}{the data pool to add the dataset wrapper to (maf_data_pool)}

\item{label}{label (string)
\enumerate{ 
\item Load the specified dataset into memory if required.
\item Update maf_data_pool (a reactiveVal) to indicate the dataset has been loaded.
\item Return the relevant maf_dataset_wrapper. See \strong{Accessing Properties} section for details.
}}
}
\value{
maf_dataset_wrapper.
}
\description{
wraps mod_select_dataset_from_maf_data_pool_pickerinput_server.
Instead of simply returning a unique_name, this function will:
}
\section{Accessing Properties (Quick Reference)}{


\tabular{rrr}{
\strong{MAF object:} \tab \tab \code{maf_dataset_wrapper()$loaded_data} \cr
\strong{unique name:} \tab \tab \code{maf_dataset_wrapper()$unique_name} \cr
\strong{short name:} \tab \tab \code{maf_dataset_wrapper()$short_name} \cr
\strong{full name:} \tab \tab \code{maf_dataset_wrapper()$display_name} \cr 
\strong{source:} \tab \tab \code{maf_dataset_wrapper()$name_of_data_source} \cr 
}


See ?new_maf_dataset_wrapper for the full list of properties
}

