% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_shinyfile_import.R
\name{mod_shinyfile_import_server}
\alias{mod_shinyfile_import_server}
\title{shinyfile_import Server Functions}
\usage{
mod_shinyfile_import_server(id)
}
\value{
chosen filepath (string) (reactive). If multiple = TRUE, it returns a character vector instead of a string. When no file is selected, returns character(0)
}
\description{
shinyfile_import Server Functions
}
