% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_shinyfile_import.R
\name{mod_shinyfile_import_ui}
\alias{mod_shinyfile_import_ui}
\title{shinyfile_import UI Function}
\usage{
mod_shinyfile_import_ui(
  id,
  title,
  label,
  multiple = FALSE,
  tooltip_text = "",
  tooltip_placement = "right",
  trigger = "hover"
)
}
\arguments{
\item{id}{Internal parameters for {shiny}.}

\item{title}{The content of the tooltip.}

\item{label}{Display label for the control, or \code{NULL} for no label.}

\item{multiple}{Whether the user should be allowed to select and upload
multiple files at once. \strong{Does not work on older browsers, including
Internet Explorer 9 and earlier.}}

\item{tooltip_text}{The content of the tooltip (string / HTML)}

\item{tooltip_placement}{Where the tooltip should appear relative to its target (top, bottom, left, or right). Defaults to "right"}

\item{trigger}{What action should cause the tooltip to appear? (hover, focus, click, or manual). Defaults to "hover".}
}
\description{
Adds a shinyFilesButton
}
