% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_shinyfiles_get_clinical_featurefile_path.R
\name{mod_shinyfiles_get_clinical_featurefile_path_server}
\alias{mod_shinyfiles_get_clinical_featurefile_path_server}
\title{Get Path to Clinical Feature File}
\usage{
mod_shinyfiles_get_clinical_featurefile_path_server(id)
}
\arguments{
\item{id}{Internal parameters for {shiny}.}
}
\value{
chosen filepath (string) (reactive). If multiple = TRUE, it returns a character vector instead of a string. When no file is selected, returns character(0)
}
\description{
Server code for getting path to clinical feature file using shinyfiles button.
No QC done here. Make sure file is valid when creating MAF object from maf + clinical feature file
}
