% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_shinyfiles_get_clinical_featurefile_path.R
\name{mod_shinyfiles_get_clinical_featurefile_path_ui}
\alias{mod_shinyfiles_get_clinical_featurefile_path_ui}
\title{shinyfiles_get_clinical_featurefile_path UI Function}
\usage{
mod_shinyfiles_get_clinical_featurefile_path_ui(id)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}
}
\description{
A shiny Module.
}
