% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_shinyfiles_get_maf_path.R
\name{mod_shinyfiles_get_maf_path_server}
\alias{mod_shinyfiles_get_maf_path_server}
\title{shinyfiles_get_maf_path Server Functions}
\usage{
mod_shinyfiles_get_maf_path_server(id)
}
\value{
chosen filepath (string) (reactive). If multiple = TRUE, it returns a character vector instead of a string. When no file is selected, returns character(0)
}
\description{
shinyfiles_get_maf_path Server Functions
}
