% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_single_cohort_summary.R
\name{mod_single_cohort_summary_tables_and_plots_server}
\alias{mod_single_cohort_summary_tables_and_plots_server}
\title{wrapper for running multiple mod_single_cohort_summary_tables_server modules}
\usage{
mod_single_cohort_summary_tables_and_plots_server(id, maf, cohortName)
}
\arguments{
\item{id}{}

\item{maf}{maf object (reactive)}

\item{cohortName}{cohort Name (string, reactive)}
}
\value{

}
\description{
wrapper for running multiple mod_single_cohort_summary_tables_server modules
}
