% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_single_cohort_summary.R
\name{mod_single_cohort_summary_tables_and_plots_ui}
\alias{mod_single_cohort_summary_tables_and_plots_ui}
\title{wrapper for running multiple mod_single_cohort_summary_tables_ui modules}
\usage{
mod_single_cohort_summary_tables_and_plots_ui(id)
}
\arguments{
\item{id}{Internal parameters for {shiny}.}
}
\description{
wrapper for running multiple mod_single_cohort_summary_tables_ui modules
}
