% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_compare_cohorts.R
\name{moduleCompareCohortsServer}
\alias{moduleCompareCohortsServer}
\title{moduleCompareCohortsServer: Serverside code for comparing}
\usage{
moduleCompareCohortsServer(id, maf_data_pool)
}
\arguments{
\item{id}{}

\item{maf}{}

\item{id_compare_settings}{List returned by moduleCompareSettingsServer.}
}
\value{

}
\description{
moduleCompareCohortsServer: Serverside code for comparing
}
