% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_compare_settings.R
\name{moduleCompareSettingsServer}
\alias{moduleCompareSettingsServer}
\title{moduleCompareSettingsServer}
\usage{
moduleCompareSettingsServer(id)
}
\arguments{
\item{id}{}
}
\value{
named list containing cohort comparison settings. 
minmut: the minimum number samples (in any cohort) which must carry a mutation in a gene for it to be included in analysis. '
usecnv: whether or not to include cnv data in fischer test if available.
}
\description{
moduleCompareSettingsServer
}
