% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_download_plot.R
\name{moduleDownloadPlotServer}
\alias{moduleDownloadPlotServer}
\title{Saves Plots as SVG.}
\usage{
moduleDownloadPlotServer(
  id,
  session_parent,
  plotOutputId,
  plotting_function,
  default_filename = "download"
)
}
\arguments{
\item{id}{Module ID. Must be identical to moduleDownloadPlotUI}

\item{plotOutputId}{"ID of the rendered plot"}

\item{plotting_function}{"function that when run with no arguments will create the plot. Can make by wrapping the plot call in its own function. e.g. for plot(mtcars) you could do:  plotting_function <- reactive ({ function() {plot(mtcars)} }) then pass 'plotting_function()'"}

\item{default_filename}{defualt basename of downloaded file}

\item{parent}{Session object of the calling module (used to extract plot details)}
}
\value{

}
\description{
Saves Plots as SVG.
}
