% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_download_plot.R
\name{moduleDownloadPlotUI}
\alias{moduleDownloadPlotUI}
\title{Creates a shinyWidgets::downloadbttn}
\usage{
moduleDownloadPlotUI(
  id,
  circle = FALSE,
  label = "Download",
  style = "unite",
  color = "default",
  size = "default",
  status = "default",
  icon = NULL,
  tooltip_placement = "right",
  tooltip_text = "",
  right = FALSE,
  up = FALSE,
  width = "200px",
  margin = "10px",
  inline = FALSE,
  ...
)
}
\arguments{
\item{id}{Module ID}

\item{circle}{Logical. Use a circle button}

\item{label}{The label that should appear on the button.}

\item{style}{Style of the button, to choose between simple, bordered, minimal, stretch, jelly, gradient, fill, material-circle, material-flat, pill, float, unite.}

\item{color}{Color of the button : default, primary, warning, danger, success, royal.}

\item{size}{Size of the button : xs,sm, md, lg.}

\item{status}{Add a class to the buttons, you can use Bootstrap status like 'info', 'primary', 'danger', 'warning' or 'success'.
Or use an arbitrary strings to add a custom class, e.g. : with \code{status = 'myClass'}, buttons will have class \code{btn-myClass}.}

\item{icon}{An icon to appear on the button.}

\item{tooltip_placement}{Where the tooltip should appear relative to its target (top, bottom, left, or right). Defaults to "right" (string)}

\item{tooltip_text}{Tooltip text (string)}

\item{right}{Logical. The dropdown menu starts on the right.}

\item{up}{Logical. Display the dropdown menu above.}

\item{width}{Width of the dropdown menu content.}

\item{margin}{Value of the dropdown margin-right and margin-left menu content.}

\item{inline}{use an inline (\code{span()}) or block container (\code{div()}) for the output.}

\item{...}{List of tag to be displayed into the dropdown menu.}
}
\value{
Nothing. Function run for its side effects
}
\description{
Creates a shinyWidgets::downloadbttn
}
