% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_utility_merge_mafs.R
\name{moduleMergeMafServer}
\alias{moduleMergeMafServer}
\title{moduleMergeMafServer}
\usage{
moduleMergeMafServer(id, maf_data_pool)
}
\arguments{
\item{id}{}
}
\value{
merged maf
}
\description{
moduleMergeMafServer
}
