% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_plot_lollipop2.R
\name{modulePlotLollipop2UI}
\alias{modulePlotLollipop2UI}
\title{plotforest: Wraps around maftools::forestPlot to more explicitly handle pval VS fdr option and allow both to be passed while onl using the value set by bool 'threshold_on_fdr'}
\usage{
modulePlotLollipop2UI(id)
}
\arguments{
\item{mafCompareRes}{}

\item{pVal}{}

\item{fdr}{}

\item{color}{}

\item{geneFontSize}{}

\item{lineWidth}{}

\item{titleSize}{}
}
\value{

}
\description{
plotforest: Wraps around maftools::forestPlot to more explicitly handle pval VS fdr option and allow both to be passed while onl using the value set by bool 'threshold_on_fdr'
}
