% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_import_maf.R
\name{moduleReadMafServer}
\alias{moduleReadMafServer}
\title{Read MAF server function}
\usage{
moduleReadMafServer(id)
}
\arguments{
\item{id}{Internal parameters for {shiny}.}
}
\value{
named list containing 'maf' (type maf), 'cohortName' (type string), 'clinicalData' (type df), 'clinicalDataSupplied' (bool),
}
\description{
Read MAF server function
}
