% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_import_maf.R
\name{moduleReadMafUI}
\alias{moduleReadMafUI}
\title{Title}
\usage{
moduleReadMafUI(
  id,
  panel_name,
  cohort_name,
  label_cohort_name = "name",
  label_maf = "maf",
  label_clinical_data = "clinical features"
)
}
\arguments{
\item{id}{links ui and server components of module}

\item{panel_name}{"panel name (required)"}

\item{cohort_name}{"panel name (required)"}

\item{label_cohort_name}{"name of the cohort (required)"}

\item{label_maf}{label of maf input field ["maf"]}

\item{label_clinical_data}{label of clinicalData input field ["clinical data"]}

\item{label_name}{"label of the cohort name field"}
}
\value{

}
\description{
Title
}
