% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_maf_data_pool.R
\name{new_maf_data_pool}
\alias{new_maf_data_pool}
\title{Constructor Objects of Class: \strong{maf_data_pool}}
\usage{
new_maf_data_pool()
}
\value{
object of class maf_data_pool (maf_data_pool)
}
\description{
The maf_data_pool class is simply a list of maf_dataset_wrapper objects.
}
\examples{
maf_data_pool <- new_maf_data_pool()
}
