% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_maf_dataset_wrapper.R
\name{new_maf_dataset_wrapper}
\alias{new_maf_dataset_wrapper}
\title{Constructor Objects of Class: \strong{maf_dataset_wrapper}}
\usage{
new_maf_dataset_wrapper(
  maf_data_pool,
  display_name,
  short_name,
  unique_name,
  start_status,
  data_description,
  is_dataset_downloadable,
  function_to_download_data = function() {
     return(NA)
 },
  is_dataset_loadable = TRUE,
  function_to_load_data,
  name_of_data_source = "unknown",
  local_path_to_data = "",
  clinical_data = NA,
  datatype_of_stored_object = "",
  derived_from = NA,
  loaded_data = NA,
  rnaseq_filepath = NA,
  number_of_samples = NA
)
}
\arguments{
\item{maf_data_pool}{A maf_data_pool object. Used to check supplied 'unique_name' is actually going to be unique. If not, characaters are appended to make it truly unique in the context of the supplied dataframe.  (maf_data_pool)}

\item{display_name}{Name that the end-user will see (string)}

\item{short_name}{Abbreviated dataset name (string)}

\item{unique_name}{Some unique identifier. (string)}

\item{start_status}{one of: "not_downloaded","not_loaded", "ready" (string)}

\item{data_description}{a description of the dataset (string)}

\item{is_dataset_downloadable}{is the dataset downloadable? The alternative is if its packaged in a compressed form with the tool.}

\item{function_to_download_data}{a function that when run will download the required data return the path to which it was downloaded if successful. Will return NA if it failed (function that returns string or NA). If is_dataset_downloadable is false, this is ignored (just use default)}

\item{is_dataset_loadable}{is dataset loadable (flag)}

\item{function_to_load_data}{a function that takes 'local_path_to_data' as its argument and returns the loaded maf object when complete. 
The function needs to take a single argument, but it doesn't actually have to use it. 
For example, when loading tcga data using the TCGAmutations package, this function could be \strong{function(filepath){maftools::tcgaLoad()}}. 
This function completely ignores the filepath argument but you NEED to include it anyway (function)}

\item{name_of_data_source}{name of the data source. For Example "USER" or "TCGA" or "PCAWG" (string)}

\item{local_path_to_data}{a path to data which will be configured based on function_to_download_data. If this option is configured ahead of time (string)}

\item{clinical_data}{the clinical data of the object}

\item{datatype_of_stored_object}{type of stored data object. Not used for anything right now, just interesting metadata. examples are \*.Rds or \*.mafs (string)}

\item{derived_from}{the maf_dataset_wrapper_object from which the new object was derived. If the dataset was obtained directly from an online source, leave as NA (maf_dataset_wrapper or NA)}

\item{loaded_data}{the loaded R object (MAF or NA)}

\item{rnaseq_filepath}{the path to the rnaseq data (string)}

\item{number_of_samples}{the number of samples in the cohort (integer)}
}
\value{
an object of type maf_dataset_wrapper (maf_dataset_wrapper)
}
\description{
maf_dataset_wrapper objects store the details of each dataset including the functions to download/load the data and its current status
}
\section{All Properties}{


\itemize{
\item display_name
\item short_name
\item unique_name
\item status
\item data_description
\item download_data
\item load_data
\item name_of_data_source
\item local_path_to_data
\item datatype_of_stored_object
\item loaded_data
\item derived_from
\item rnaseq_filepath
\item number_of_samples
}
}

