% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcawg_mutations.R
\name{pcawg_available}
\alias{pcawg_available}
\title{List Available PCAWG datasets}
\usage{
pcawg_available(
 
    file_listing_projects = system.file("extdata/pcawg_no_intergenic_by_project_code_rds/pcawg_projects.tsv",
    package = "shinymaftools")
)
}
\arguments{
\item{file_listing_projects}{path to a tsv (w/ header) listing all possible projects in the first column}
}
\value{
a dataframe of the available datasets where the first column contains the values you'd need to feed into pcawg_load
}
\description{
List Available PCAWG datasets
}
\examples{
pcawg_available()
}
