% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wrapping_pcawg.R
\name{pcawg_dataset_to_data_pool}
\alias{pcawg_dataset_to_data_pool}
\title{Add PCAWG dataset to a data pool}
\usage{
pcawg_dataset_to_data_pool(maf_data_pool, pcawg_study_abbreviation)
}
\arguments{
\item{maf_data_pool}{the datapool to add pcawg study too (maf_data_pool)}

\item{pcawg_study_abbreviation}{Name of PCAWG study. Use pcawg_available to see options (string)}
}
\value{
returns the data pool with the specified dataset added (maf_data_pool)
}
\description{
Add PCAWG dataset to a data pool
}
\examples{
pcawg_dataset_to_data_pool(new_maf_data_pool(), "Bone-Cart")

}
