% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wrapping_pcawg.R
\name{pcawg_dataset_to_maf_dataset_wrapper}
\alias{pcawg_dataset_to_maf_dataset_wrapper}
\title{Create dataset wrapper from pcawg dataset}
\usage{
pcawg_dataset_to_maf_dataset_wrapper(pcawg_study_abbreviation)
}
\arguments{
\item{pcawg_study_abbreviation}{Name of PCAWG study. Use pcawg_available to see options (string)}
}
\value{
maf_dataset_wrapper object
}
\description{
Create dataset wrapper from pcawg dataset
}
\examples{
pcawg_dataset_to_maf_dataset_wrapper("Bone-Cart")
}
