% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcawg_mutations.R
\name{pcawg_load}
\alias{pcawg_load}
\title{Load a PCAWG dataset into Memory}
\usage{
pcawg_load(
  dataset_name,
 
    directory_containing_projectfiles = system.file("extdata/pcawg_no_intergenic_by_project_code_rds",
    package = "shinymaftools")
)
}
\arguments{
\item{dataset_name}{a valid pcawg dataset name. For a list of available dataset names, run pcawg_available() (string)}

\item{directory_containing_projectfiles}{the directory containing Rds files storing the maf objects. Each file must have be named 'dataset_name'.rds. (string)}
}
\value{
a maf object
}
\description{
Loads specified PCAWG dataset into memory (as a maf object).
}
\examples{
pcawg_load("Biliary-AdenoCA")
}
