% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_plot_rainforest.R
\name{plotforest}
\alias{plotforest}
\title{plotforest: Wraps around maftools::forestPlot to more explicitly handle pval VS fdr option and allow both to be passed while onl using the value set by bool 'threshold_on_fdr'}
\usage{
plotforest(
  mafCompareRes,
  pVal = 0.05,
  fdr = 0.05,
  threshold_on_fdr,
  color = NULL,
  geneFontSize = 1.2,
  lineWidth = 2.2,
  titleSize = 1.2
)
}
\arguments{
\item{mafCompareRes}{results from \code{\link[maftools]{mafCompare}}}

\item{pVal}{p-value threshold. Default 0.05.}

\item{fdr}{fdr threshold. Default NULL. If provided uses adjusted pvalues (fdr).}

\item{threshold_on_fdr}{should we threshold on FDR (TRUE) or on pValue (false) (flag)}

\item{color}{vector of two colors for the lines. Default 'maroon' and 'royalblue'}

\item{geneFontSize}{Font size for gene symbols. Default 0.8}

\item{lineWidth}{line width for CI bars. Default 1}

\item{titleSize}{font size for titles. Default 1.2}
}
\description{
plotforest: Wraps around maftools::forestPlot to more explicitly handle pval VS fdr option and allow both to be passed while onl using the value set by bool 'threshold_on_fdr'
}
