% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_data_pool_table.R
\name{shinyInput}
\alias{shinyInput}
\title{Programmatically create a Shiny input}
\usage{
shinyInput(FUN, n, id, label, ...)
}
\arguments{
\item{FUN}{function to create the input}

\item{n}{number of inputs to be created}

\item{id}{ID prefix for each input}
}
\description{
Programmatically create a Shiny input
}
