% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_TCGAmutations_load_with_typed_metadata.R
\name{TCGAmutations_load_with_typed_metadata}
\alias{TCGAmutations_load_with_typed_metadata}
\title{Load TCGA data}
\usage{
TCGAmutations_load_with_typed_metadata(study, source = "Firehose")
}
\arguments{
\item{study}{Study names to load. Use  \code{\link[maftools]{tcgaAvailable}} to see available options.}

\item{source}{Source for MAF files. Can be \code{MC3} or \code{Firehose}. Default \code{MC3}. Argument may be abbreviated (M or F)}
}
\value{
MAF object with correctly typed clinical.data dataframe (MAF)
}
\description{
Load TCGA data with correctly typed clinical data. 
By default, maftools::tcgaLoad makes all Firehose dataset clinical features character vectors
}
