% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wrapping_user_imports.R
\name{user_data_filepath_to_class_maf_dataset_wrapper}
\alias{user_data_filepath_to_class_maf_dataset_wrapper}
\title{Title}
\usage{
user_data_filepath_to_class_maf_dataset_wrapper(
  filepath,
  clinicalData = NA,
  maf_data_pool,
  display_name,
  short_name,
  description = "User specified file",
  data_source = "USER",
  loaded_data = NA
)
}
\arguments{
\item{filepath}{tab delimited MAF file. File can also be gz compressed. Required. Alternatively, you can also provide already read MAF file as a dataframe.}

\item{clinicalData}{Clinical data associated with each sample/Tumor_Sample_Barcode in MAF. Could be a text file or a data.frame. Default NULL.}

\item{maf_data_pool}{A maf_data_pool object. Used to check supplied 'unique_name' is actually going to be unique. If not, characaters are appended to make it truly unique in the context of the supplied dataframe.  (maf_data_pool)}

\item{display_name}{Name that the end-user will see (string)}

\item{short_name}{Abbreviated dataset name (string)}

\item{loaded_data}{the loaded R object (MAF or NA)}
}
\value{
maf_dataset_wrapper
}
\description{
Title
}
\seealso{
Other DataToWrapper: 
\code{\link{tcga_dataset_to_data_pool}()},
\code{\link{tcga_dataset_to_maf_dataset_wrapper}()},
\code{\link{tcga_datasets_to_data_pool}()},
\code{\link{user_to_dataset_to_data_pool}()}
}
\concept{DataToWrapper}
