% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wrapping_user_imports.R
\name{user_to_dataset_to_data_pool}
\alias{user_to_dataset_to_data_pool}
\title{Add User Data to data pool}
\usage{
user_to_dataset_to_data_pool(
  maf_data_pool,
  filepath,
  clinicalData = NA,
  display_name,
  short_name,
  description = "User specified file",
  data_source = "USER",
  loaded_data = NA
)
}
\arguments{
\item{maf_data_pool}{the data pool to add the dataset wrapper to (maf_data_pool)}

\item{filepath}{path to maf file}

\item{clinicalData}{Clinical data associated with each sample/Tumor_Sample_Barcode in MAF. Could be a text file or a data.frame. Default NULL.}

\item{display_name}{Name that the end-user will see (string)}

\item{short_name}{Abbreviated dataset name (string)}

\item{description}{dataset description}

\item{data_source}{data source}

\item{loaded_data}{the loaded R object (MAF or NA)}
}
\value{
returns a data pool object with extra dataset added (maf_data_pool)
}
\description{
Takes a filepath to user dataset + some medata and creates a maf_dataset_wrapper object, then adds the wrapper to the maf_data_pool
}
\seealso{
Other DataToWrapper: 
\code{\link{tcga_dataset_to_data_pool}()},
\code{\link{tcga_dataset_to_maf_dataset_wrapper}()},
\code{\link{tcga_datasets_to_data_pool}()},
\code{\link{user_data_filepath_to_class_maf_dataset_wrapper}()}
}
\concept{DataToWrapper}
