% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wrapping_zero.R
\name{zero_datasets_to_data_pool}
\alias{zero_datasets_to_data_pool}
\title{Adds all ZERO cohort to data pool}
\usage{
zero_datasets_to_data_pool(maf_data_pool)
}
\arguments{
\item{maf_data_pool}{the data pool to add the dataset wrapper to (maf_data_pool)}
}
\value{
returns a data pool object with extra dataset added (maf_data_pool)
}
\description{
Takes a maf_data_poolobject and adds maf_dataset_wrappers for all ZERO datasets available
}
\examples{
zero_datasets_to_data_pool(new_maf_data_pool())
}
\seealso{
Other DataToWrapper: 
\code{\link{tcga_dataset_to_data_pool}()},
\code{\link{tcga_dataset_to_maf_dataset_wrapper}()},
\code{\link{tcga_datasets_to_data_pool}()},
\code{\link{user_data_filepath_to_class_maf_dataset_wrapper}()},
\code{\link{user_to_dataset_to_data_pool}()},
\code{\link{zero_dataset_to_data_pool}()},
\code{\link{zero_dataset_to_maf_dataset_wrapper}()}
}
\concept{DataToWrapper}
