% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crux-package.R
\docType{package}
\name{crux}
\alias{crux}
\alias{crux-package}
\title{Identify the Crux of an Article}
\description{
Methods are provided to retrieve HTML content and return extracted
metadata and summarised plain text. Further methods are provided to classify
URLs with or without making network calls. Based on \url{https://github.com/chimbori/crux}.
}
\details{
Based on \url{https://github.com/chimbori/crux}.
\itemize{
\item URL: \url{https://gitlab.com/hrbrmstr/crux}
\item BugReports: \url{https://gitlab.com/hrbrmstr/crux/issues}
\item Javadoc: \url{https://www.javadoc.io/doc/com.chimbori.crux/crux/2.0.2}
}
}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
