% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_external_tools_load_dataframe.R
\name{external_tools_load_bbglab_cgi}
\alias{external_tools_load_bbglab_cgi}
\title{Load tool metadata into environment. Returns}
\usage{
external_tools_load_bbglab_cgi(external_tools_df = data.frame())
}
\arguments{
\item{external_tools_df}{the dataframe to add tool metadata to. By default, will create and return a new dataframe}
}
\value{
external_tools_df with metadata of tool appeneded (data.frame)
}
\description{
Appends  metadata for the "Cancer Genome Interpreter (cgi) tool onto external_tools_df
}
\examples{
CRUX:::external_tools_load_bbglab_cgi()
}
