% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_external_tools_load_dataframe.R
\name{external_tools_load_proteinpaint}
\alias{external_tools_load_proteinpaint}
\title{Load tool metadata into global variable}
\usage{
external_tools_load_proteinpaint(external_tools_df = data.frame())
}
\arguments{
\item{external_tools_df}{the dataframe to add tool metadata to. By default, will create and return a new dataframe}
}
\value{
external_tools_df with metadata of tool appeneded (data.frame)
}
\description{
Appends  metadata for the OncodriveCLUSTL tool onto external_tools_df
}
