% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_maftools_extending_functions.R
\name{maftools_chrom_23_and_24_to_X_and_Y}
\alias{maftools_chrom_23_and_24_to_X_and_Y}
\title{Change chrom 23 and 24 to x and y in maf object. This will convert}
\usage{
maftools_chrom_23_and_24_to_X_and_Y(maf)
}
\arguments{
\item{maf}{}
}
\value{

}
\description{
Change chrom 23 and 24 to x and y in maf object. This will convert
}
