% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_import_clinical_featurefile.R
\name{mod_import_clinical_featurefile_server}
\alias{mod_import_clinical_featurefile_server}
\title{import_clinical_featurefile Server Functions}
\usage{
mod_import_clinical_featurefile_server(id, maf_path)
}
\arguments{
\item{maf}{a reactive maf object which the supplied clinical feature file is associated with}
}
\value{
the object with metadata from a clinical feature
}
\description{
Imports a clinical feature file.
}
