% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_shinyfile_import.R
\name{mod_shinyfile_import_ui}
\alias{mod_shinyfile_import_ui}
\title{shinyfile_import UI Function}
\usage{
mod_shinyfile_import_ui(
  id,
  title,
  label,
  multiple = FALSE,
  buttonType = "dark",
  style = NULL,
  viewtype = "detail",
  tooltip_text = "",
  tooltip_placement = "right",
  trigger = "hover"
)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}

\item{tooltip_text}{The content of the tooltip (string / HTML)}

\item{tooltip_placement}{Where the tooltip should appear relative to its target (top, bottom, left, or right). Defaults to "right"}

\item{trigger}{What action should cause the tooltip to appear? (hover, focus, click, or manual). Defaults to "hover".}
}
\description{
Adds a shinyFilesButton
}
