% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_split_maf_by_clinical_feature.R
\name{moduleSubsetByClinicalFeaturesServer}
\alias{moduleSubsetByClinicalFeaturesServer}
\title{moduleSubsetByClinicalFeaturesServer}
\usage{
moduleSubsetByClinicalFeaturesServer(id, maf, clinicalData)
}
\arguments{
\item{id}{Internal parameters for {shiny}.}

\item{maf}{reactive maf object}

\item{clinicalData}{clinical data object (reactive)}
}
\value{
named list: 'maf1' (type maf), 'maf2' (type string),
}
\description{
moduleSubsetByClinicalFeaturesServer
}
