% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_plot_rainforest.R
\name{plotforest}
\alias{plotforest}
\title{plotforest: Wraps around maftools::forestPlot to more explicitly handle pval VS fdr option and allow both to be passed while onl using the value set by bool 'threshold_on_fdr'}
\usage{
plotforest(
  mafCompareRes,
  pVal = 0.05,
  fdr = 0.05,
  threshold_on_fdr,
  color = NULL,
  geneFontSize = 1.2,
  lineWidth = 2.2,
  titleSize = 1.2
)
}
\arguments{
\item{mafCompareRes}{}

\item{pVal}{}

\item{fdr}{}

\item{color}{}

\item{geneFontSize}{}

\item{lineWidth}{}

\item{titleSize}{}
}
\value{

}
\description{
plotforest: Wraps around maftools::forestPlot to more explicitly handle pval VS fdr option and allow both to be passed while onl using the value set by bool 'threshold_on_fdr'
}
