% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_filetype_guesser.R
\name{read_maf_flexible}
\alias{read_maf_flexible}
\title{Read MAF from ANNOVAR OR MAF input}
\usage{
read_maf_flexible(
  path_mutations,
  refBuild = NULL,
  path_clindata = NULL,
  filetype = c("AUTO", "ANNOVAR", "MAF")
)
}
\arguments{
\item{path_mutations}{path to a MAF / ANNOVAR file}

\item{refBuild}{which reference version is being used}
}
\value{
one of MAF, ANNOVAR, VCF, OTHER
}
\description{
Automatically detects filetype from
}
