% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wrapping_tcga.R
\name{tcga_dataset_to_data_pool}
\alias{tcga_dataset_to_data_pool}
\title{Add TCGA cohort to data pool}
\usage{
tcga_dataset_to_data_pool(
  tcga_study_abbreviation,
  maf_data_pool,
  source = "Firehose"
)
}
\arguments{
\item{tcga_study_abbreviation}{a TCGA cohort abbreviation (see maftools::tcgaAvailable()) (string)}

\item{maf_data_pool}{the data pool to add the dataset wrapper to (maf_data_pool)}

\item{source}{'MC3' or 'Firehose'. Source of TCGA data to use. See ?maftools::tcgaLoad for details (string)}
}
\value{
returns a data pool object with extra dataset added (maf_data_pool)
}
\description{
Takes a TCGA study abbreviation, creates a tcga_dataset_to_maf_dataset_wrapper object for the relevant dataset and adds the wrapper to the maf_data_pool
}
\examples{
CRUX:::tcga_dataset_to_data_pool("ACC", CRUX:::new_maf_data_pool())
}
\seealso{
Other DataToWrapper: 
\code{\link{tcga_dataset_to_maf_dataset_wrapper}()},
\code{\link{tcga_datasets_to_data_pool}()},
\code{\link{user_data_filepath_to_class_maf_dataset_wrapper}()},
\code{\link{user_to_dataset_to_data_pool}()}
}
\concept{DataToWrapper}
