% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wrapping_tcga.R
\name{tcga_dataset_to_maf_dataset_wrapper}
\alias{tcga_dataset_to_maf_dataset_wrapper}
\title{TCGA to maf_dataset_wrapper}
\usage{
tcga_dataset_to_maf_dataset_wrapper(
  maf_data_pool,
  tcga_study_abbreviation,
  source = "Firehose"
)
}
\arguments{
\item{maf_data_pool}{the data pool to add the dataset wrapper to (maf_data_pool)}

\item{tcga_study_abbreviation}{a TCGA cohort abbreviation (see TCGAmutations::tcga_available()) (string)}

\item{source}{'MC3' or 'Firehose'. Source of TCGA data to use. See ?TCGAmutations::tcga_load for details (string)}
}
\value{
functions and values associated with specified TCGA cohort (maf_dataset_wrapper)
}
\description{
Takes a TCGA cohort abbreviation
}
\examples{
ACC_maf_dataset_wrapper <- tcga_dataset_to_maf_dataset_wrapper("ACC")
}
\seealso{
Other DataToWrapper: 
\code{\link{tcga_dataset_to_data_pool}()},
\code{\link{tcga_datasets_to_data_pool}()},
\code{\link{user_data_filepath_to_class_maf_dataset_wrapper}()},
\code{\link{user_to_dataset_to_data_pool}()}
}
\concept{DataToWrapper}
