% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_TCGAmutations_load_with_typed_metadata.R
\name{TCGAmutations_load_with_typed_metadata}
\alias{TCGAmutations_load_with_typed_metadata}
\title{Load TCGA data}
\usage{
TCGAmutations_load_with_typed_metadata(study, source = "MC3")
}
\arguments{
\item{study}{Study names to load. Use  \code{\link[TCGAmutations]{tcga_available}} to see available options.}

\item{source}{Can be \code{MC3} or \code{Firehose}. Default \code{MC3}
It will be usefull if you want to operate multiple \code{MAF}s.}
}
\value{
MAF object with correctly typed clinical.data dataframe (MAF)
}
\description{
Load TCGA data with correctly typed clinical data. 
By default, TCGAmutations::tcga_load makes all MC3 dataset clinical features character vectors
}
\examples{
TCGAmutations_load_with_typed_metadata(study = "GBM")
}
